toda a parte lógica e estrutural finalizada. Iniciado as estilizações
