+++ semana9/labex/src/pages/ApplicationForm/styles.jsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+
+export const ApplyForm = styled.form`
+  display: flex;
+  flex-direction: column;
+  max-width: 350px;
+  margin: 0 auto;
+  input{
+    margin-bottom: 10px;
+    padding: 5px;
+  }
+  select{
+    padding: 5px;
+    margin-bottom: 10px;
+  }
+`
