+++ semana9/labex/src/pages/Home/styles.jsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+
+export const GeneralContainer = styled.div`
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+`
+
+export const DoubleButtons = styled.div`
+  display: flex;
+  justify-content: space-between;
+  width: 250px;
+  margin: 1rem 0;
+`
